Add KVM Switch support Disable MOUSEKEY_ENABLE, NKRO_ENABLE and enable NO_USB_STARTUP_CHECK for the KVM switch to not detect the keyboard as mouse or media device and to keep the keyboard on and responsive between switches.
Fix MacOSX DE_LESS and DE_CIRC key switch problem
Just swapped them. This is not a beautiful fix.

Add key to get to layer 0 when stuck
So far I can't fix the problem with pressing two OSL Keys in a row without releasing the first one and then getting stuck in the first layer. That's why a key was introduced to get out of this by switching back to layer 0.